+++ CHANGELOG.md
@@ -23,11 +23,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
-- modified `miactl marketplace delete` command to accept either the `objectId` or a `itemId`-`version` tuple that identifies the item to be deleted.
+- modified `miactl marketplace delete` command to accept either the `objectId` or a `companyId`-`itemId`-`version` tuple that identifies the item to be deleted.
+- modified `miactl marketplace get` command to accept either the `objectId` or a `companyId`-`itemId`-`version` tuple that identifies the item to be retrieved.
 
 ### BREAKING
 
 - `miactl marketplace delete` does not accept anymore the id as argument, it should be provided to the flag `--object-id`
+- `miactl marketplace get` does not accept anymore the id as argument, it should be provided to the flag `--object-id`
 
 ## [0.9.0] - 2023-11-15